hpcemerg: Add thread creation on init
Remove nr_thpool from param struct because we don't need to determine
the number of thread pools. Spawning only one thread is enough.

Use pthread library to create an emergency pool for long running task
when we hit the signal handler. Since inside the signal handler we're
not allowed to block, so we have to move the work to this thread pool.
Currently it does nothing, we will have a mechanism to transfer the
work to this thread pool.

Signed-off-by: Ammar Faizi <ammarfaizi2@gmail.com>