hpcemerg: scripts/build: Fix path generation in DEPFLAGS
If we use:
```
  DEPFLAGS = -MT "$@" -MMD -MP -MF "$(@:$(BASE_DIR)/%.o=$(BASE_DEP_DIR)/%.d)"
```
then the DEPFLAGS only works for files that are compiled to "*.o"
extension. We need DEPFLAGS to also work on compile that directly
generates executable code (IOW, taget file that is not "*.o").

As such, generalize the dependency generated file. Remove the "*.o".
Now we use:
```
  DEPFLAGS = -MT "$@" -MMD -MP -MF "$(@:$(BASE_DIR)/%=$(BASE_DEP_DIR)/%.d)"
```
which can work for literally every file that lives in the base
directory.

Signed-off-by: Ammar Faizi <ammarfaizi2@gmail.com>