hpcemerg: Add warn trap for x86 arch
This adds warn trap for x86 arch. There are 4 new macros here:
  - WARN()
  - WARN_ONCE()
  - WARN_ON()
  - WARN_ON_ONCE()

In order to support this, I also change the user_func function
prototype in the struct hpcemerg_init_param. We create a new
struct named hpcemerg_sig_ctx, we use this to contain all data
to be passed to the user function.
```
  void (*user_func)(struct hpcemerg_sig_ctx *sig_ctx);
```
Now the user can see more data from the signal handler.

Signed-off-by: Ammar Faizi <ammarfaizi2@gmail.com>